Preserve upstream type-handle, layout, and allocation-helper changes while retaining experimental twice-pointer alignment. Update moved compiler logic, cDAC contracts, auxiliary symbols, and Decimal128 alignment.

Validation is source-level only: the clean-head clr+libs+host baseline failed with MSVC C1083 in NativeAOT datadescriptor.cpp. Builds and tests remain blocked; this local merge is committed with that gap explicitly accepted.

+++ src/libraries/Common/src/Interop/Windows/SspiCli/Interop.SSPI.cs
@@ -538,8 +538,8 @@ internal static partial class Interop
         [LibraryImport(Interop.Libraries.SspiCli, SetLastError = true)]
         internal static partial SECURITY_STATUS SetCredentialsAttributesW(
             in CredHandle handlePtr,
-            long ulAttribute,
+            uint ulAttribute,
             in SecPkgCred_ClientCertPolicy pBuffer,
-            long cbBuffer);
+            uint cbBuffer);
     }
 }
